Graph refresh
This patch introduces a first version of graph refresh, for both full
and targeted refreshes.

The initial full refresh is performed retrieving all the objects from
the Kubernetes API. The targeted refresh is implemented processing the
Kubernetes notices.

In order to keep track of what notices have been processed already the
refresh worker uses a local JSON file where it saves the list of notices
that have already been processed.