Adapt the provider to the latest KubeVirt API
The KubeVirt team has decided to use `OfflineVirtualMachine` as the name
for the custom resource that represents virtual machines that are not
running. We assumed `StoredVirtualMachine` in the past. This patch
changes the provider to use the new name and to unify the names used for
methods and variables. They will be now like this:

  OfflineVirtualMachine -> offline_vm
  VirtualMachine -> live_vm
  Template -> template

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>