Fix parsing of node adresses
Currently the code that parses the node IP and host name doesn't work
correctly for full refresh and partial refresh, because the data
provided by the `kubeclient` gem for notices is different than the data
retrieved directly.

The previous patch has introduced a fix for that `kubeclient` issue.
This patch changes the code so that it will work for both full and
partial refreshes.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>