Adjust the memory calculator to Kubevirt 0.2.0
In Kubevirt 0.2.0 the way that virtual machine memory is specified has
changed. In previous versions it was specified like this:

  memory:
    unit: MB
    value: 64

See the example here:

  https://github.com/kubevirt/kubevirt/blob/v0.1.0/cluster/vm.yaml#L34-L36

Since version 0.2.0 it is specified like this:

  memory: 64M

See the example here:

  https://github.com/kubevirt/kubevirt/blob/v0.2.0/cluster/vm.yaml#L10

This means that the `MemoryCalculator` class needs to be changed
accordingly to support the new format.

Also the provision request and the parsing of the reported virtual
machines were updated to support the updated memory format.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Signed-off-by: Moti Asayag <masayag@redhat.com>