Add full refresh tests
This patch adds an set of tests for the full refresh. The tests create a
collector from a JSON file and then check that the parser generates the
expected inventory objects.

Currently these tests only do basic checks about clusters, storages and
hosts; they should be extended in the future to check the other objects.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>