Add VM Provision dialog
Kubevirt is a virtualization provider that allows to create virtual
machines. Currently, it is pretty limited and not fully defined,
therefore the initial dialog allows the user to specific the VM memory
only.

For vm provisioning, the :linked_clone attribute is ignored, as the
decision of disk provision type is made by the storage backend.
In addition, we don't control the node on which the vm will be
scheduled, therefore the implementation of allowed_hosts is irrelevant in this
context.