Implement auth methods
The default_auth_status is being used by the UI to detect the validity
of the provider. For kubevirt, the default auth_status refers to the
parent provider, while kubevirt auth_status is detected by the :kubevirt
auth_type.

For that purpose, there is a need to report always 'true' for the
'default_auth_status', so the kubevirt provider can be edited when being
accessed as an infra provider.